requirements.txt
fastapi==0.104.1          # Framework web rápido y moderno para crear APIs en Python (basado en ASGI).
uvicorn==0.24.0           # Servidor ASGI ligero y rápido que ejecuta aplicaciones FastAPI.
sqlalchemy==2.0.23        # ORM para manejar bases de datos relacionales como PostgreSQL desde Python.
psycopg2-binary==2.9.9    # Driver que permite conectar Python con PostgreSQL (usado por SQLAlchemy).
python-multipart==0.0.6   # Permite manejar formularios y archivos enviados en peticiones HTTP (multipart/form-data).
jinja2==3.1.2             # Motor de plantillas HTML, útil si tu FastAPI genera páginas web dinámicas.
pandas==2.1.3             # Librería para análisis y manipulación de datos (muy usada para procesar datasets).
openpyxl==3.1.2           # Permite leer y escribir archivos Excel (.xlsx), ideal si importas/exportas datos.
python-dotenv==1.0.0      # Carga variables de entorno desde un archivo .env (para configurar credenciales o rutas).

passlib==1.7.4
bcrypt==5.0.0
itsdangerous==2.2.0